app: split page runtimes
