+++ crates/pages/src/lib.rs
@@ -16,7 +16,10 @@ pub mod scroll_persist;
 pub mod search;
 pub mod server;
 pub mod settings;
+pub mod settings_actions;
 #[cfg(not(target_os = "android"))]
 pub mod theme_editor;
 #[cfg(all(not(target_arch = "wasm32"), not(target_os = "android")))]
 pub mod ytdlp;
+#[cfg(all(not(target_arch = "wasm32"), not(target_os = "android")))]
+pub mod ytdlp_jobs;
